Release v0.4.0: Linux support, setup wizard, health check, zero warnings
- Sync all version strings to 0.4.0
- Cross-platform memory detection (macOS sysctl / Linux /proc/meminfo)
- Linux llama-server paths (/usr/bin, /snap/bin) and install hints
- Fix sed -i syntax for GNU sed on Linux
- Add setup wizard shown on first launch for LLM configuration
- Guard agent mode toggle behind LLM config check
- Add LLM server health check with auto-restart on crash
- Add React error boundary and toast notification system
- Add Linux bundle config (deb, AppImage) in tauri.conf.json
- Add GitHub Actions workflows for macOS and Linux desktop releases
- Fix all 26 clippy warnings across all crates
- Fix test files missing React import and stale version strings

+++ crates/localcode-core/src/indexing/embeddings.rs
@@ -92,7 +92,7 @@ pub fn bm25_score(
     let mut score: f32 = 0.0;
     for term in &query_terms {
         let tf = *tf_map.get(term.as_str()).unwrap_or(&0) as f32;
-        let df = *doc_freqs.get(term.as_str()).unwrap_or(&0) as usize;
+        let df = *doc_freqs.get(term.as_str()).unwrap_or(&0);
 
         if tf == 0.0 || df == 0 {
             continue;
